+++ src/components/Wizard/Wizard.tsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import FormWizard from 'react-form-wizard-component';
 import { FormWizardMethods } from 'react-form-wizard-component/dist/types/types/FormWizard';
 import { useIntl } from 'react-intl';
@@ -9,7 +9,7 @@ import { DetailsSVGIcon } from '../../common/icons/DetailsSVGIcon';
 import { SelectionSVGIcon } from '../../common/icons/SelectionSVGIcon';
 import { UpdateSVGIcon } from '../../common/icons/UpdateSVGIcon';
 import { useI18n } from '../../i18n/I18nProvider';
-import { SnackbarManager } from '../common/SnackBar/SnackbarManager';
+import { SnackbarManager } from '../common/Snackbar/SnackbarManager';
 import { Step } from '../common/Step/Step';
 import { ISummary } from '../common/Tree/hooks/treeCatalogData.hook';
 import { MetadataConfirm } from './MetadataConfirm/MetadataConfirm';
@@ -26,15 +26,24 @@ export const Wizard: React.FC = () => {
   const [selectedItem, setSelectedItem] = useState<CatalogTreeNode | undefined>(undefined);
   const [itemsSummary, setItemsSummary] = useState<ISummary | undefined>(undefined);
   const [disabled, setDisabled] = useState<boolean>(true);
+  const [shouldSubmit, setShouldSubmit] = useState<boolean>(false);
+  const [isCompleted, setIsCompleted] = useState<boolean>(false);
   const intl = useIntl();
-  const wizardRef = useRef<FormWizardMethods>(null);
   const { locale } = useI18n();
+  const wizardRef = useRef<FormWizardMethods>(null);
+
+  useEffect(() => {
+    if (isCompleted) {
+      setCatalogTree(undefined);
+      setSelectedItem(undefined);
+      setDisabled(true);
+      setShouldSubmit(false);
+      wizardRef.current?.goToTab(0);
+    }
+  }, [isCompleted]);
   
-  const handleComplete = () => {
-    setCatalogTree(undefined);
-    setSelectedItem(undefined);
-    setDisabled(true);
-    wizardRef.current?.goToTab(0);
+  const handleComplete = async () => {
+    setShouldSubmit(true);
   };
 
   return (
@@ -100,6 +109,9 @@ export const Wizard: React.FC = () => {
               <MetadataConfirm
                 setIsNextBtnDisabled={(val) => { setDisabled(val) }}
                 selectedItem={selectedItem}
+                shouldSubmit={shouldSubmit}
+                setShouldSubmit={setShouldSubmit}
+                setIsCompleted={setIsCompleted}
               />
             }
           </Step>