+++ src/utils/requestHandler.ts
@@ -1,5 +1,7 @@
 import axios, { AxiosRequestConfig, AxiosResponse, Method } from 'axios';
+import { SnackbarManager } from '../components/common/Snackbar/SnackbarManager';
 import appConfig from './Config';
+import { getSnackbarErrorMessage } from './snackbarError';
 
 interface IResource {
   url: string;
@@ -60,4 +62,30 @@ export const requestHandlerWithToken = async (url: string, method: string, param
   }
 
   return requestHandler(url, method, reqConfig);
+};
+
+export const execute = async (
+  url: string,
+  method: 'GET' | 'POST' | 'DELETE',
+  data?: Record<string, unknown>,
+  submitErrorToSnackbarQueue = true
+): Promise<Record<string, unknown>[] | string | undefined> => {
+  try {
+    const response = await requestExecutor(
+      {
+        url,
+        injectToken: true
+      },
+      method,
+      {
+        ...(data ?? {})
+      }
+    );
+    return response?.status === 204 ? 'OK' : response?.data;
+  } catch (error) {
+    if (submitErrorToSnackbarQueue) {
+      SnackbarManager.notify(getSnackbarErrorMessage((error as any).message as string));
+    }
+    throw error;
+  }
 };